move save_thresholds function
